requirements.txt
dvc
mlflow==2.2.2
notebook
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
tensorflow==2.12.0
gdown
-e .